melhorei util de testes
