+++ ufabc_planner-backend/jest.config.ts
@@ -4,7 +4,7 @@ import { pathsToModuleNameMapper } from 'ts-jest';
 const config: Config.InitialOptions = {
   preset: 'ts-jest',
   testTimeout: 30000,
-  modulePaths: ['<rootDir>/../src/'],
+  modulePaths: ['<rootDir>/src/'],
   testEnvironment: 'node',
   rootDir: '.',
   globals: {
