+++ ufabc_planner-backend/test/constants.ts
@@ -1,4 +1,4 @@
-export const validPassword = '12345';
+export const validPassword = '123';
 
 export const randomNames = ['Eliza', 'Martinho', 'Felipe', 'Gabriel', 'Jorge', 'Vinicius', 'Leonardo', 'Lucas'];
 export const randomEmails = ['yahoo.com', 'gmail.com', 'hotmail.com', 'fake.com', 'user.com', 'test.com', 'mayo.com'];
