--- ufabc_planner-backend/test/utils/database.ts
@@ -1,9 +1,0 @@
-import { prisma } from 'utils/prisma';
-
-export async function closeConnections() {
-  const deleteUser = prisma.user.deleteMany();
-
-  await prisma.$transaction([deleteUser]);
-
-  await prisma.$disconnect();
-}
